feat: move header components to header folder
