+++ src/components/Header/index.tsx
@@ -1,7 +1,7 @@
 import { ShoppingCart } from 'phosphor-react'
 import { Link, useParams } from 'react-router-dom'
-import { HeaderSectionPerPage } from '../HeaderSectionPerPage'
-import { MarkHeadPhone } from '../MarkHeadPhone'
+import { HeaderSectionPerPage } from './components/HeaderSectionPerPage'
+import { MarkHeadPhone } from './components/MarkHeadPhone'
 import {
   Container,
   ContainerIcon,
@@ -25,6 +25,7 @@ export function Header () {
               <DropdownMenuTrigger>
                 <img src="/assets/Group.png" alt="" />
               </DropdownMenuTrigger>
+
               <DropdownMenuContent>
                 <ul>
                   <li>
@@ -42,10 +43,12 @@ export function Header () {
                 </ul>
               </DropdownMenuContent>
             </DropdownMenuRoot>
+
             <Link to="/">
               <img src="/assets/audiophile.png" alt="Logo esrito audiophile" />
             </Link>
           </ContainerImage>
+
           <ContainerLink>
             <li>
               <Link to={'/'}>Home</Link>
@@ -60,6 +63,7 @@ export function Header () {
               <Link to={'/products/earphones'}>Earphones</Link>
             </li>
           </ContainerLink>
+
           <ContainerIcon>
             <ShoppingCart size={24} />
           </ContainerIcon>
