--- src/components/HeaderSectionPerPage/index.tsx
@@ -1,13 +1,0 @@
-import { Container } from './styles'
-
-interface HeaderSectionPerPageProps {
-  params: string;
-}
-
-export function HeaderSectionPerPage ({ params }: HeaderSectionPerPageProps) {
-  return (
-    <Container>
-      <h1>{params}</h1>
-    </Container>
-  )
-}
